[codegen] factor: dual __co__ support.
- rewrite the streams to allow generating code
  for a compilation with multiple __co__
  functions.
- Fix: lit config to always remove host comment
  to avoid FileCheck failures. In addition, 9
  existing but not detected test failures are
  fixed.
- Use '--native-f16' to support native f16 type.
- Use Regex to match compilation options, where
  corresponding test failures are fixed.
- Reduce compiler warning in building choreo.

+++ codegen_cuda.cpp
@@ -854,10 +854,11 @@ bool CUDACodeGen::Visit(AST::Call& c) {
   if (c.template_params != nullptr) {
     fs << "<";
     bool need_delimiter = false;
-    for (int i = 0; i < c.template_params->Count(); ++i) {
+    for (size_t i = 0; i < c.template_params->Count(); ++i) {
       if (need_delimiter) fs << ", ";
       need_delimiter = true;
-      fs << STR(cast<AST::Expr>(c.template_params->ValueAt(i))->compile_time_signature);
+      fs << STR(cast<AST::Expr>(c.template_params->ValueAt(i))
+                    ->compile_time_signature);
     }
     fs << ">";
   }
